Split indexController into userController
